Simplify controllers by adding Store.$listen
One of the main complexities in the controller code was the necessity
to get back into the Angular world from the Reflux world by calling
`$scope.$apply`.  This change adds a `$listen` method to all stores
which automatically does this for you.  This change, along with some
other simplifications (e.g., refreshing the scope during *any* event,
not just `CHANGE` events) reduces most of the "listen" functions to a
single line.