Use `listenables` in Reflux Stores
This simplifies things in several ways.  First, many stores no
longer need an `init` method at all since all they did was manually
register actions.  Second, it allows a mixin to discover what actions
its target responds to, and to discover their child actions. This
eliminates the need for the target of a mixin to provide some mixin-
specific way of doing this.